tests/requirements.txt
numpy
protobuf==3.20.3
ffmpeg-python
git+https://github.com/diggerdu/visqol-py.git@452eb5c4f17fd2404f968ec2eeadfcad74925485
